добавляет отображение даты начала и конца маршрута
